+++ app/modules/kit-library/admin-menu-items/editor-one-website-templates-menu.php
@@ -2,7 +2,7 @@
 
 namespace Elementor\App\Modules\KitLibrary\AdminMenuItems;
 
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 use Elementor\Plugin;
