+++ core/admin/editor-one-menu/interfaces/menu-item-interface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Elementor\Modules\EditorOne\Classes\Menu;
+namespace Elementor\Core\Admin\EditorOneMenu\Interfaces;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -22,3 +22,4 @@ interface Menu_Item_Interface {
 
 	public function get_group_id();
 }
+
