+++ core/admin/editor-one-menu/interfaces/menu-item-third-level-interface.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace Elementor\Modules\EditorOne\Classes\Menu;
+namespace Elementor\Core\Admin\EditorOneMenu\Interfaces;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
-
 
 interface Menu_Item_Third_Level_Interface extends Menu_Item_Interface {
 
@@ -13,3 +12,4 @@ interface Menu_Item_Third_Level_Interface extends Menu_Item_Interface {
 
 	public function has_children(): bool;
 }
+
