+++ core/admin/editor-one-menu/menu/abstract-level4-menu-item.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Elementor\Modules\EditorOne\Classes\Menu\Items;
+namespace Elementor\Core\Admin\EditorOneMenu\Menu;
 
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -31,3 +31,4 @@ abstract class Abstract_Level4_Menu_Item implements Menu_Item_Interface {
 
 	abstract public function get_group_id();
 }
+
