+++ core/admin/editor-one-menu/menu/abstract-menu-item.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Elementor\Modules\EditorOne\Classes\Menu\Items;
+namespace Elementor\Core\Admin\EditorOneMenu\Menu;
 
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Third_Level_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Third_Level_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -39,3 +39,4 @@ abstract class Abstract_Menu_Item implements Menu_Item_Third_Level_Interface {
 
 	abstract public function get_icon(): string;
 }
+
