+++ core/admin/editor-one-menu/menu/base-third-party-menu-item.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Elementor\Core\Admin\EditorOneMenu;
+namespace Elementor\Core\Admin\EditorOneMenu\Menu;
 
 use Elementor\Core\Admin\EditorOneMenu\Interfaces\Third_Party_Menu_Item;
 
@@ -9,13 +9,8 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 abstract class Base_Third_Party_Menu_Item implements Third_Party_Menu_Item {
-
 	public function is_visible(): bool {
 		return current_user_can( $this->get_capability() );
-	}
-
-	public function get_icon(): string {
-		return 'settings';
 	}
 
 	abstract public function get_id(): string;
@@ -28,3 +23,4 @@ abstract class Base_Third_Party_Menu_Item implements Third_Party_Menu_Item {
 
 	abstract public function get_owner(): string;
 }
+
