+++ core/admin/editor-one-menu/menu/legacy-submenu-item-not-mapped.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Elementor\Modules\EditorOne\Classes\Menu\Items;
+namespace Elementor\Core\Admin\EditorOneMenu\Menu;
 
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Third_Level_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Third_Level_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -63,3 +63,4 @@ class Legacy_Submenu_Item_Not_Mapped implements Menu_Item_Third_Level_Interface 
 		return false;
 	}
 }
+
