+++ core/common/modules/connect/admin-menu-items/editor-one-connect-menu.php
@@ -4,7 +4,7 @@ namespace Elementor\Core\Common\Modules\Connect\AdminMenuItems;
 
 use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item_With_Page;
 use Elementor\Core\Common\Modules\Connect\Apps\Base_App;
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 use Elementor\Plugin;
