+++ modules/editor-one/module.php
@@ -3,12 +3,10 @@
 namespace Elementor\Modules\EditorOne;
 
 use Elementor\Core\Admin\Admin;
-use Elementor\Core\Admin\EditorOneMenu\Base_Third_Party_Menu_Item;
-use Elementor\Core\Admin\EditorOneMenu\Third_Party_Menu_Manager;
 use Elementor\Core\Base\Module as BaseModule;
 use Elementor\Core\Experiments\Manager as Experiments_Manager;
 use Elementor\Modules\EditorOne\Classes\Editor_One_Pointer;
-use Elementor\Modules\EditorOne\Components\Elementor_One_Menu_Manager;
+use Elementor\Core\Admin\EditorOneMenu\Elementor_One_Menu_Manager;
 use Elementor\Modules\EditorOne\Components\Sidebar_Navigation_Handler;
 use Elementor\Plugin;
