+++ modules/element-manager/admin-menu-items/editor-one-elements-manager-menu.php
@@ -3,7 +3,7 @@
 namespace Elementor\Modules\ElementManager\AdminMenuItems;
 
 use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item_With_Page;
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 use Elementor\Modules\ElementManager\Module;
