+++ modules/floating-buttons/admin-menu-items/editor-one-floating-elements-menu.php
@@ -2,7 +2,7 @@
 
 namespace Elementor\Modules\FloatingButtons\AdminMenuItems;
 
-use Elementor\Modules\EditorOne\Classes\Menu\Menu_Item_Interface;
+use Elementor\Core\Admin\EditorOneMenu\Interfaces\Menu_Item_Interface;
 use Elementor\Modules\EditorOne\Classes\Menu_Config;
 use Elementor\Modules\FloatingButtons\Module;
